+++ docx-core/src/documents/elements/run.rs
@@ -194,6 +194,13 @@ impl Run {
         self
     }
 
+    pub fn add_tc(mut self, tc: InstrTC) -> Run {
+        self = self.add_field_char(crate::types::FieldCharType::Begin, false);
+        self = self.add_instr_text(InstrText::TC(tc));
+        self = self.add_field_char(crate::types::FieldCharType::End, false);
+        self
+    }
+
     pub fn add_instr_text(mut self, i: InstrText) -> Run {
         self.children.push(RunChild::InstrText(Box::new(i)));
         self
@@ -256,6 +263,11 @@ impl Run {
         self
     }
 
+    pub fn stretch(mut self, v: i32) -> Run {
+        self.run_property = self.run_property.stretch(v);
+        self
+    }
+
     pub fn color(mut self, color: impl Into<String>) -> Run {
         self.run_property = self.run_property.color(color);
         self
@@ -283,6 +295,11 @@ impl Run {
 
     pub fn strike(mut self) -> Run {
         self.run_property = self.run_property.strike();
+        self
+    }
+
+    pub fn dstrike(mut self) -> Run {
+        self.run_property = self.run_property.dstrike();
         self
     }