+++ docx-core/src/reader/run_property.rs
@@ -92,6 +92,13 @@ impl ElementReader for RunProperty {
                             }
                             rp = rp.strike();
                         }
+                        XMLElement::Dstrike => {
+                            if !read_bool(&attributes) {
+                                rp.dstrike = Some(Dstrike::new().disable());
+                                continue;
+                            }
+                            rp = rp.dstrike();
+                        }
                         XMLElement::VertAlign => {
                             if let Ok(v) = VertAlignType::from_str(&attributes[0].value) {
                                 rp = rp.vert_align(v)
@@ -120,6 +127,11 @@ impl ElementReader for RunProperty {
                                 continue;
                             }
                             rp = rp.italic();
+                        }
+                        XMLElement::Shading => {
+                            if let Ok(shd) = Shading::read(r, &attributes) {
+                                rp = rp.shading(shd);
+                            }
                         }
                         XMLElement::Vanish => rp = rp.vanish(),
                         XMLElement::SpecVanish => rp = rp.spec_vanish(),